// wellwise_backend/src/modules/user/health/dto/update-health-profile.dto.ts
export class UpdateHealthProfileDto {
  dateOfBirth?: string;
  gender?: string;
  weightKg?: number;
  heightCm?: number;
  lifestyle?: string;
  sleepQuality?: number;
  stressLevel?: number;
  focusLevel?: number;
  physicalActivity?: string;
  nutritionHabits?: string;
  waterIntake?: number;
  notes?: string;
}